Moorcheh/requirements.txt
# requirements.txt - GoFish Moorcheh Brain

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Environment variables
python-dotenv>=0.21.0

# HTTP client for Moorcheh API
requests>=2.31.0

# Data validation
pydantic>=2.0.0

# Template rendering
jinja2>=3.1.0

# File upload support
python-multipart>=0.0.6
